feat(#752): remove old implementation of MaxStack
